fix(encoders): break latent adapter import cycle, repair moved-symbol scripts
The adapter grab-bag split (2e91ad8) made the encoder modules import the
new adapter modules at top level, while the adapter modules import
`src.r2dreamer.encoders.constants` at top level. That forms a genuine
cycle between the `encoders` and `adapters` package `__init__`s; it only
avoids crashing today because `vggt_adapter` transitively pulls in
`encoders.cnn` early, which is fragile and order-dependent.

Revert the encoders->adapters edges back to lazy `importlib.import_module`
calls inside each `_build_adapter_for_extractor` (the pre-split pattern),
in encoders/__init__.py, house_points_pose.py and house_global_embedding.py.
This keeps the valuable file split (house_context_adapter / house_points_adapter
/ token_adapters) while removing the module-level cycle, so both
`import src.r2dreamer.adapters` and `import src.r2dreamer.encoders` succeed
regardless of order.

Also repoint 4 scripts at the new home of the moved symbols
(VGGTHousePointsPoseObsAdapter and _vggt_output_field now live in
adapters/house_points_adapter.py, not hybrid_adapter.py).

Verified on CPU: both import orders succeed and the encoder suite
(test_encoders, test_house_global_embedding_adapter, test_agent,
test_dims_consistency) passes 66 / 2 skipped.
