+++ src/r2dreamer/encoders/__init__.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from collections.abc import Mapping
+from importlib import import_module
 from types import MappingProxyType
 from typing import TYPE_CHECKING, Any
 
@@ -30,12 +31,6 @@ from src.r2dreamer.encoders.house_points_pose import (
 )
 from src.r2dreamer.encoders.pointnet import PointNetHousePointsPoseEncoder
 from src.r2dreamer.encoders.transformer import TokenTransformerEncoder
-from src.r2dreamer.adapters.house_context_adapter import VGGTHouseContextObsAdapter
-from src.r2dreamer.adapters.hybrid_adapter import HybridObsAdapter
-from src.r2dreamer.adapters.token_adapters import (
-    VGGTHouseFullTokenObsAdapter,
-    VGGTHouseGlobalTokenObsAdapter,
-)
 
 if TYPE_CHECKING:
     from src.r2dreamer.adapters.obs_adapter import ObsAdapter
@@ -60,7 +55,8 @@ class HybridEncoder(VGGTEncoder):
     variant_key = "hybrid"
 
     def _build_adapter_for_extractor(self, extractor) -> ObsAdapter:
-        return HybridObsAdapter(
+        adapter_module = import_module("src.r2dreamer.adapters.hybrid_adapter")
+        return adapter_module.HybridObsAdapter(
             extractor,
             env_render_resolution=self.env_render_resolution,
             encoder_module_cls=self.module_cls,
@@ -145,7 +141,8 @@ class VGGTHouseContextEncoder(VGGTEncoder):
         return super().new_adapter()
 
     def _build_adapter_for_extractor(self, extractor) -> ObsAdapter:
-        return VGGTHouseContextObsAdapter(
+        adapter_module = import_module("src.r2dreamer.adapters.house_context_adapter")
+        return adapter_module.VGGTHouseContextObsAdapter(
             extractor,
             context_transformer=self._context_transformer,
             static_house_context_path=self._static_house_context_path,
@@ -158,7 +155,8 @@ class VGGTHouseFullTokenNoGateEncoder(VGGTHouseContextEncoder):
     variant_key = "vggt_house_full_tokens_nogate"
 
     def _build_adapter_for_extractor(self, extractor) -> ObsAdapter:
-        return VGGTHouseFullTokenObsAdapter(extractor)
+        adapter_module = import_module("src.r2dreamer.adapters.token_adapters")
+        return adapter_module.VGGTHouseFullTokenObsAdapter(extractor)
 
 
 class VGGTHouseGlobalTokenNoGateEncoder(VGGTHouseContextEncoder):
@@ -167,4 +165,5 @@ class VGGTHouseGlobalTokenNoGateEncoder(VGGTHouseContextEncoder):
     variant_key = "vggt_house_global_tokens_nogate"
 
     def _build_adapter_for_extractor(self, extractor) -> ObsAdapter:
-        return VGGTHouseGlobalTokenObsAdapter(extractor)
+        adapter_module = import_module("src.r2dreamer.adapters.token_adapters")
+        return adapter_module.VGGTHouseGlobalTokenObsAdapter(extractor)
